+++ src/scheduler/check.cpp
@@ -77,30 +77,30 @@
 #include <pbs_internal.h>
 #include <log.h>
 #include <libutil.h>
-#include "check.hpp"
-#include "config.hpp"
-#include "server_info.hpp"
-#include "queue_info.hpp"
-#include "job_info.hpp"
-#include "misc.hpp"
-#include "constant.hpp"
-#include "globals.hpp"
-#include "dedtime.hpp"
-#include "node_info.hpp"
-#include "fifo.hpp"
-#include "resource_resv.hpp"
+#include "check.h"
+#include "config.h"
+#include "server_info.h"
+#include "queue_info.h"
+#include "job_info.h"
+#include "misc.h"
+#include "constant.h"
+#include "globals.h"
+#include "dedtime.h"
+#include "node_info.h"
+#include "fifo.h"
+#include "resource_resv.h"
 #ifdef NAS
-#include "site_code.hpp"
+#include "site_code.h"
 #endif
-#include "node_partition.hpp"
-#include "sort.hpp"
-#include "server_info.hpp"
-#include "queue_info.hpp"
+#include "node_partition.h"
+#include "sort.h"
+#include "server_info.h"
+#include "queue_info.h"
 #include "limits_if.h"
-#include "simulate.hpp"
-#include "resource.hpp"
-#include "buckets.hpp"
-#include "pbs_bitmap.hpp"
+#include "simulate.h"
+#include "resource.h"
+#include "buckets.h"
+#include "pbs_bitmap.h"
 
 
 /**
