+++ src/scheduler/fifo.cpp
@@ -64,40 +64,40 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
-#include "fifo.hpp"
-#include "queue_info.hpp"
-#include "server_info.hpp"
-#include "node_info.hpp"
-#include "check.hpp"
-#include "constant.hpp"
-#include "job_info.hpp"
-#include "misc.hpp"
-#include "config.hpp"
-#include "sort.hpp"
-#include "parse.hpp"
-#include "globals.hpp"
-#include "prev_job_info.hpp"
-#include "fairshare.hpp"
-#include "prime.hpp"
-#include "dedtime.hpp"
-#include "resv_info.hpp"
+#include "fifo.h"
+#include "queue_info.h"
+#include "server_info.h"
+#include "node_info.h"
+#include "check.h"
+#include "constant.h"
+#include "job_info.h"
+#include "misc.h"
+#include "config.h"
+#include "sort.h"
+#include "parse.h"
+#include "globals.h"
+#include "prev_job_info.h"
+#include "fairshare.h"
+#include "prime.h"
+#include "dedtime.h"
+#include "resv_info.h"
 #include "range.h"
-#include "resource_resv.hpp"
-#include "simulate.hpp"
-#include "node_partition.hpp"
-#include "resource.hpp"
-#include "resource_resv.hpp"
+#include "resource_resv.h"
+#include "simulate.h"
+#include "node_partition.h"
+#include "resource.h"
+#include "resource_resv.h"
 #include "pbs_share.h"
 #include "pbs_internal.h"
 #include "limits_if.h"
 #include "pbs_version.h"
-#include "buckets.hpp"
-#include "multi_threading.hpp"
+#include "buckets.h"
+#include "multi_threading.h"
 #include "pbs_python.h"
 #include "libpbs.h"
 
 #ifdef NAS
-#include "site_code.hpp"
+#include "site_code.h"
 #endif
 
 /* a list of running jobs from the last scheduling cycle */
