+++ src/scheduler/get_4byte.cpp
@@ -45,8 +45,8 @@
 #include <stdlib.h>
 #include "dis.h"
 #include "sched_cmds.h"
-#include "data_types.hpp"
-#include "fifo.hpp"
+#include "data_types.h"
+#include "fifo.h"
 
 #include <sys/types.h>
 #include <sys/time.h>
