+++ src/scheduler/node_info.cpp
@@ -132,27 +132,27 @@
 #include <libutil.h>
 #include <pbs_internal.h>
 #include "attribute.h"
-#include "node_info.hpp"
-#include "server_info.hpp"
-#include "job_info.hpp"
-#include "misc.hpp"
-#include "globals.hpp"
-#include "check.hpp"
-#include "constant.hpp"
-#include "config.hpp"
-#include "resource_resv.hpp"
-#include "simulate.hpp"
-#include "sort.hpp"
-#include "node_partition.hpp"
-#include "resource.hpp"
+#include "node_info.h"
+#include "server_info.h"
+#include "job_info.h"
+#include "misc.h"
+#include "globals.h"
+#include "check.h"
+#include "constant.h"
+#include "config.h"
+#include "resource_resv.h"
+#include "simulate.h"
+#include "sort.h"
+#include "node_partition.h"
+#include "resource.h"
 #include "pbs_internal.h"
-#include "server_info.hpp"
+#include "server_info.h"
 #include "pbs_share.h"
-#include "pbs_bitmap.hpp"
+#include "pbs_bitmap.h"
 #include "pbs_license.h"
-#include "multi_threading.hpp"
+#include "multi_threading.h"
 #ifdef NAS
-#include "site_code.hpp"
+#include "site_code.h"
 #endif
