+++ src/scheduler/node_info.h
@@ -43,7 +43,7 @@
 extern "C" {
 #endif
 
-#include "data_types.hpp"
+#include "data_types.h"
 #include <pbs_ifl.h>
 
 void query_node_info_chunk(th_data_query_ninfo *data);
