+++ src/scheduler/pbs_sched.cpp
@@ -74,14 +74,14 @@
 #include <sys/resource.h>
 
 #include "auth.h"
-#include "config.hpp"
-#include "fifo.hpp"
-#include "globals.hpp"
+#include "config.h"
+#include "fifo.h"
+#include "globals.h"
 #include "libpbs.h"
 #include "libsec.h"
 #include "list_link.h"
 #include "log.h"
-#include "multi_threading.hpp"
+#include "multi_threading.h"
 #include "net_connect.h"
 #include "pbs_ecl.h"
 #include "pbs_error.h"
@@ -91,7 +91,7 @@
 #include "pbs_version.h"
 #include "portability.h"
 #include "rm.h"
-#include "misc.hpp"
+#include "misc.h"
 #include "sched_cmds.h"
 #include "server_limits.h"
 #include "tpp.h"
