+++ src/scheduler/resource_resv.cpp
@@ -103,22 +103,22 @@
 #include <pthread.h>
 #include <libutil.h>
 #include "pbs_config.h"
-#include "data_types.hpp"
-#include "resource_resv.hpp"
-#include "job_info.hpp"
-#include "resv_info.hpp"
-#include "node_info.hpp"
-#include "misc.hpp"
-#include "node_partition.hpp"
-#include "constant.hpp"
-#include "globals.hpp"
-#include "resource.hpp"
+#include "data_types.h"
+#include "resource_resv.h"
+#include "job_info.h"
+#include "resv_info.h"
+#include "node_info.h"
+#include "misc.h"
+#include "node_partition.h"
+#include "constant.h"
+#include "globals.h"
+#include "resource.h"
 #include "pbs_internal.h"
-#include "check.hpp"
-#include "fifo.hpp"
+#include "check.h"
+#include "fifo.h"
 #include "range.h"
-#include "simulate.hpp"
-#include "multi_threading.hpp"
+#include "simulate.h"
+#include "multi_threading.h"
 
 
 /**
