+++ src/scheduler/resv_info.cpp
@@ -70,22 +70,22 @@
 #include <log.h>
 #include <libutil.h>
 
-#include "data_types.hpp"
-#include "resv_info.hpp"
-#include "job_info.hpp"
-#include "queue_info.hpp"
-#include "server_info.hpp"
-#include "misc.hpp"
-#include "sort.hpp"
-#include "globals.hpp"
-#include "node_info.hpp"
-#include "resource_resv.hpp"
-#include "resource.hpp"
-#include "fifo.hpp"
-#include "check.hpp"
-#include "simulate.hpp"
-#include "constant.hpp"
-#include "node_partition.hpp"
+#include "data_types.h"
+#include "resv_info.h"
+#include "job_info.h"
+#include "queue_info.h"
+#include "server_info.h"
+#include "misc.h"
+#include "sort.h"
+#include "globals.h"
+#include "node_info.h"
+#include "resource_resv.h"
+#include "resource.h"
+#include "fifo.h"
+#include "check.h"
+#include "simulate.h"
+#include "constant.h"
+#include "node_partition.h"
 #include "pbs_internal.h"
 
 /**
