+++ src/scheduler/server_info.cpp
@@ -124,34 +124,34 @@
 #include "pbs_error.h"
 #include "log.h"
 #include "pbs_share.h"
-#include "server_info.hpp"
-#include "constant.hpp"
-#include "queue_info.hpp"
-#include "job_info.hpp"
-#include "misc.hpp"
-#include "config.hpp"
-#include "node_info.hpp"
-#include "globals.hpp"
-#include "resv_info.hpp"
-#include "sort.hpp"
-#include "resource_resv.hpp"
-#include "state_count.hpp"
-#include "node_partition.hpp"
-#include "resource.hpp"
+#include "server_info.h"
+#include "constant.h"
+#include "queue_info.h"
+#include "job_info.h"
+#include "misc.h"
+#include "config.h"
+#include "node_info.h"
+#include "globals.h"
+#include "resv_info.h"
+#include "sort.h"
+#include "resource_resv.h"
+#include "state_count.h"
+#include "node_partition.h"
+#include "resource.h"
 #include "assert.h"
 #include "limits_if.h"
-#include "resource.hpp"
+#include "resource.h"
 #include "pbs_internal.h"
-#include "simulate.hpp"
-#include "fairshare.hpp"
-#include "check.hpp"
+#include "simulate.h"
+#include "fairshare.h"
+#include "check.h"
 // #include "pbs_sched.h"
-#include "fifo.hpp"
-#include "buckets.hpp"
-#include "parse.hpp"
+#include "fifo.h"
+#include "buckets.h"
+#include "parse.h"
 #include "hook.h"
 #ifdef NAS
-#include "site_code.hpp"
+#include "site_code.h"
 #endif
 
 extern char **environ;
