+++ src/scheduler/site_code.cpp
@@ -58,33 +58,33 @@
 #include <limits.h>
 #include <assert.h>
 #include "log.h"
-#include "data_types.hpp"
-#include "fifo.hpp"
-#include "queue_info.hpp"
-#include "server_info.hpp"
-#include "node_info.hpp"
-#include "check.hpp"
-#include "constant.hpp"
-#include "job_info.hpp"
-#include "misc.hpp"
-#include "config.hpp"
-#include "sort.hpp"
-#include "parse.hpp"
-#include "globals.hpp"
-#include "prev_job_info.hpp"
-#include "fairshare.hpp"
-#include "prime.hpp"
-#include "dedtime.hpp"
-#include "resv_info.hpp"
+#include "data_types.h"
+#include "fifo.h"
+#include "queue_info.h"
+#include "server_info.h"
+#include "node_info.h"
+#include "check.h"
+#include "constant.h"
+#include "job_info.h"
+#include "misc.h"
+#include "config.h"
+#include "sort.h"
+#include "parse.h"
+#include "globals.h"
+#include "prev_job_info.h"
+#include "fairshare.h"
+#include "prime.h"
+#include "dedtime.h"
+#include "resv_info.h"
 #include "range.h"
-#include "resource.hpp"
-#include "resource_resv.hpp"
-#include "simulate.hpp"
+#include "resource.h"
+#include "resource_resv.h"
+#include "simulate.h"
 
 #ifdef NAS
 
-#include "site_data.hpp"
-#include "site_code.hpp"
+#include "site_data.h"
+#include "site_code.h"
 #include "site_queue.h"
 
 #define TJ_COST_MAX	10.0	/* Max CPU to spend searching for top jobs */
